change to hash map for mesh data holding

+++ src/ui/camera.cpp
@@ -92,7 +92,7 @@ void Camera::updateInput() {
     }
 }
 
-void Camera::updateInput(Scene& scene, size_t idx) {
+void Camera::updateInput(Scene& scene, int idx) {
     if (m_userInteraction) {
         glm::vec3 right = glm::normalize(glm::cross(m_forward, m_up)) * m_renderConfig.moveSpeed;
         glm::vec3 forward = m_forward * m_renderConfig.moveSpeed;
